Split CharacterOrder to its own class file
